# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AuthResponse:
    description: Ответ с токеном доступа и информацией о пользователе
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/dto.UserInfo'
    type: object
...
  dto.LoginRequest:
    description: Запрос на аутентификацию пользователя
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterRequest:
    description: Запрос на регистрацию нового пользователя
    properties:
      confirm_password:
        example: password123
        type: string
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        maxLength: 50
        minLength: 2
        type: string
      last_name:
        example: Doe
        maxLength: 50
        minLength: 2
        type: string
      password:
        example: password123
        maxLength: 100
        minLength: 8
        type: string
    required:
    - confirm_password
    - email
    - first_name
    - last_name
    - password
    type: object
  dto.UserInfo:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 1
        type: integer
      last_name:
        example: Doe
        type: string
    type: object
...
host: localhost:9091
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для управления логистическими операциями включая аутентификацию,
    заказы, доставки и управление складом, построенный на базе нескольких микросервисов
    с использованием gRPC и REST.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Logistics Management API
  version: "1.0"
paths:
  /auth/logout:
    post:
      description: Выполняет выход пользователя и удаляет refresh token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Выход из системы
      tags:
      - auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Выполняет вход пользователя и возвращает токены
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Некорректные данные
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Неверные учетные данные
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Аутентификация пользователя
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешная регистрация
          schema:
            properties:
              email:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              user_id:
                format: int64
                type: integer
            type: object
        "400":
          description: Некорректные данные
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Введите 'Bearer ' followed by your JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"